pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "synthdatalab"
version = "0.1.0"
description = "Industrial-Grade Synthetic Data Pipeline with Anti-Collapse Guarantees"
readme = "dataset_builder/README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "ollama>=0.3.0",
    "jsonschema>=4.17.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "numpy>=1.24.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.9",
    "types-jsonschema",
]
image = [
    "Pillow>=9.0.0",
    "pytesseract>=0.3.10",
]

[project.scripts]
synthdatalab = "dataset_builder.main:cli"

# ── Ruff (linter + formatter) ─────────────────────────────────────────────────

[tool.ruff]
target-version = "py310"
line-length = 100
exclude = [
    "__pycache__",
    ".git",
    "*.egg-info",
    "data/",
    ".venv",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "SIM", # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument (common in Click)
    "N818",   # Exception naming (not enforced here)
]

[tool.ruff.lint.isort]
known-first-party = ["config", "ingestion", "generation", "validation", "filtering", "evaluation", "schema", "analysis", "prompts"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

# ── Mypy ──────────────────────────────────────────────────────────────────────

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
no_strict_optional = true
explicit_package_bases = true
warn_unused_ignores = true
warn_return_any = false
exclude = ["data/", "tests/"]

# ── Pytest ────────────────────────────────────────────────────────────────────

[tool.pytest.ini_options]
testpaths = ["dataset_builder/tests"]
addopts = "-v --tb=short"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["dataset_builder"]
omit = [
    "dataset_builder/tests/*",
    "dataset_builder/data/*",
    "**/__pycache__/**",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "@overload",
]
